:recycle: Rename scanner components p2
