+++ src/components/shared/Scanner/CameraAccess.js
