+++ src/components/shared/Scanner/Scanner.styles.js
